refactoring for assets
